Refactor sample app to single activity
