+++ app/src/main/AndroidManifest.xml
@@ -25,15 +25,6 @@
             </intent-filter>
         </activity>
 
-        <activity
-            android:name=".GPFormActivity"
-            android:exported="true"
-            android:theme="@style/Theme.T3GamePaySDKCoreUI" />
-
-        <activity
-            android:name=".GPSavedCardActivity"
-            android:exported="true"
-            android:theme="@style/Theme.T3GamePaySDKCoreUI" />
     </application>
 
 </manifest>
