+++ app/src/main/res/layout/fragment_input_fields.xml
@@ -3,10 +3,23 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:layout_marginTop="16dp"
-    android:paddingHorizontal="@dimen/gp_form_inset_left"
-    android:paddingVertical="@dimen/gp_form_inset_top">
+    android:orientation="vertical">
+
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="?attr/colorPrimary"
+        android:title="Input Fields"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar" />
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:layout_marginTop="16dp"
+        android:paddingHorizontal="@dimen/gp_form_inset_left"
+        android:paddingVertical="@dimen/gp_form_inset_top">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -104,4 +117,5 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp" />
 
+    </LinearLayout>
 </LinearLayout>
